requirements.txt
djangorestframework>=3.11.0,<3.12.0
Django>=3.0.0,<3.1.0
flake8>=3.7.9,<3.8.0
psycopg2>=2.8.5,<2.9.0
Pillow>=7.1.2,<7.2.0